+++ api/test/trace/propagation/http_text_format_test.cc
@@ -50,7 +50,7 @@ TEST(TextMapPropagatorTest, NoSendEmptyTraceState)
   TextMapCarrierTest carrier;
   carrier.headers_ = {{"traceparent", "00-4bf92f3577b34da6a3ce929d0e0e4736-0102030405060708-01"}};
   context::Context ctx1 = context::Context{
-      "current-span",
+      trace::kSpanKey,
       nostd::shared_ptr<trace::Span>(new trace::DefaultSpan(trace::SpanContext::GetInvalid()))};
   context::Context ctx2 = format.Extract(carrier, ctx1);
   TextMapCarrierTest carrier2;
@@ -65,7 +65,7 @@ TEST(TextMapPropagatorTest, PropogateTraceState)
   carrier.headers_ = {{"traceparent", "00-4bf92f3577b34da6a3ce929d0e0e4736-0102030405060708-01"},
                       {"tracestate", "congo=t61rcWkgMzE"}};
   context::Context ctx1 = context::Context{
-      "current-span",
+      trace::kSpanKey,
       nostd::shared_ptr<trace::Span>(new trace::DefaultSpan(trace::SpanContext::GetInvalid()))};
   context::Context ctx2 = format.Extract(carrier, ctx1);
 
@@ -82,7 +82,7 @@ TEST(TextMapPropagatorTest, PropagateInvalidContext)
   // Do not propagate invalid trace context.
   TextMapCarrierTest carrier;
   context::Context ctx{
-      "current-span",
+      trace::kSpanKey,
       nostd::shared_ptr<trace::Span>(new trace::DefaultSpan(trace::SpanContext::GetInvalid()))};
   format.Inject(carrier, ctx);
   EXPECT_TRUE(carrier.headers_.count("traceparent") == 0);
@@ -105,6 +105,25 @@ TEST(TextMapPropagatorTest, SetRemoteSpan)
   EXPECT_EQ(Hex(span->GetContext().span_id()), "0102030405060708");
   EXPECT_EQ(span->GetContext().IsSampled(), true);
   EXPECT_EQ(span->GetContext().IsRemote(), true);
+}
+
+TEST(TextMapPropagatorTest, DoNotOverwriteContextWithInvalidSpan)
+{
+  TextMapCarrierTest carrier;
+  constexpr uint8_t buf_span[]  = {1, 2, 3, 4, 5, 6, 7, 8};
+  constexpr uint8_t buf_trace[] = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16};
+  trace::SpanContext span_context{trace::TraceId{buf_trace}, trace::SpanId{buf_span},
+                                  trace::TraceFlags{true}, false};
+  nostd::shared_ptr<trace::Span> sp{new trace::DefaultSpan{span_context}};
+
+  // Make sure this invalid span does not overwrite the active span context
+  carrier.headers_ = {{"traceparent", "00-FOO92f3577b34da6a3ce929d0e0e4736-010BAR0405060708-01"}};
+  context::Context ctx1{trace::kSpanKey, sp};
+  context::Context ctx2 = format.Extract(carrier, ctx1);
+  auto ctx2_span        = ctx2.GetValue(trace::kSpanKey);
+  auto span             = nostd::get<nostd::shared_ptr<trace::Span>>(ctx2_span);
+
+  EXPECT_EQ(Hex(span->GetContext().trace_id()), "0102030405060708090a0b0c0d0e0f10");
 }
 
 TEST(TextMapPropagatorTest, GetCurrentSpan)
@@ -165,7 +184,7 @@ TEST(GlobalTextMapPropagator, NoOpPropagator)
   carrier.headers_ = {{"traceparent", "00-4bf92f3577b34da6a3ce929d0e0e4736-0102030405060708-01"},
                       {"tracestate", "congo=t61rcWkgMzE"}};
   context::Context ctx1 = context::Context{
-      "current-span",
+      trace::kSpanKey,
       nostd::shared_ptr<trace::Span>(new trace::DefaultSpan(trace::SpanContext::GetInvalid()))};
   context::Context ctx2 = propagator->Extract(carrier, ctx1);
 
@@ -189,7 +208,7 @@ TEST(GlobalPropagator, SetAndGet)
   carrier.headers_ = {{"traceparent", "00-4bf92f3577b34da6a3ce929d0e0e4736-0102030405060708-01"},
                       {"tracestate", trace_state_value}};
   context::Context ctx1 = context::Context{
-      "current-span",
+      trace::kSpanKey,
       nostd::shared_ptr<trace::Span>(new trace::DefaultSpan(trace::SpanContext::GetInvalid()))};
   context::Context ctx2 = propagator->Extract(carrier, ctx1);